+++ codegen/src/api/mod.rs
@@ -122,7 +122,7 @@ pub(crate) fn get_schema_name(maybe_ref: Option<&str>) -> Option<String> {
     let r = maybe_ref?;
     let schema_name = r.strip_prefix("#/components/schemas/");
     if schema_name.is_none() {
-        tracing::warn!(
+        tracing::error!(
             component_ref = r,
             "missing #/components/schemas/ prefix on component ref"
         );